requirements.txt
Flask
gunicorn
ultralytics
Pillow
numpy
opencv-python-headless